# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header",
      "description": "Authentication token"
    }
  },
  "paths": {
    "/search": {
      "get": {
        "tags": [
          "Search Text"
        ],
        "summary": "Search by places using opencagedata service",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Specifies the query string",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/convert": {
      "post": {
        "tags": [
          "Convert ETH to USDT"
        ],
        "summary": "Convert ETH to USDT with Wallet Address, Private Key and Gas Limit",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "Convert ETH to USDT",
            "in": "body",
            "description": "Request Body for Converting ETH to USDT",
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "wallAddr": {
                  "type": "string"
                },
                "privKey": {
                  "type": "string"
                },
                "gasLimit": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/encrypt": {
      "post": {
        "tags": [
          "Encrypt Wallet Address & Private Key"
        ],
        "summary": "Encrypt Wallet Address & Private Key",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "Encryption",
            "in": "body",
            "description": "Request Body for Encryption",
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "wallAddr": {
                  "type": "string"
                },
                "privKey": {
                  "type": "string"
                },
                "gasLimit": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/decrypt": {
      "post": {
        "tags": [
          "Decrypt Wallet Address & Private Key"
        ],
        "summary": "Decrypt Wallet Address & Private Key",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "Decryption",
            "in": "body",
            "description": "Request Body for Decryption",
            "type": "string",
            "schema": {
              "type": "object",
              "properties": {
                "wallAddr": {
                  "type": "string"
                },
                "privKey": {
                  "type": "string"
                },
                "gasLimit": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    },
    "/keywordSearch": {
      "get": {
        "tags": [
          "Search Keyword"
        ],
        "summary": "Search by places using opencagedata service",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          }
        },
        "content": {
          "application/json": {}
        },
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "description": "Specifies the query string",
            "required": true,
            "type": "string"
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ]
      }
    }
  }
}